Cleanup GraphQL
