+++ src/Client.AzureDevOps/AzureDevOpsBuilderExtensions.cs
@@ -3,14 +3,13 @@ using System.Collections.Generic;
 using System.Text;
 using Magnet.Client.AzureDevOps;
 
-namespace Magnet.Client
+namespace Magnet.Client;
+
+public static class AzureDevOpsBuilderExtensions
 {
-    public static class AzureDevOpsBuilderExtensions
+    public static MagnetClientBuilder RegisterAzureDevOps(this MagnetClientBuilder builder)
     {
-        public static MagnetClientBuilder RegisterAzureDevOps(this MagnetClientBuilder builder)
-        {
-            builder.AddMessageType("WorkItem", new WorkItemMessageMapper());
-            return builder;
-        }
+        builder.AddMessageType("WorkItem", new WorkItemMessageMapper());
+        return builder;
     }
 }
