+++ src/Client.AzureDevOps/Client.AzureDevOps.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AssemblyName>Magnet.Client.AzureDevOps</AssemblyName>
     <RootNamespace>Magnet.Client.AzureDevOps</RootNamespace>
     <PackageId>Magnet.Client.AzureDevOps</PackageId>
