+++ src/Client.AzureDevOps/WorkItemMessageMapper.cs
@@ -1,15 +1,14 @@
 using Magnet.Client.Mappers;
 using Newtonsoft.Json;
 
-namespace Magnet.Client.AzureDevOps
+namespace Magnet.Client.AzureDevOps;
+
+public class WorkItemMessageMapper : IMessageMapper<WorkItemEventMessage>
 {
-    public class WorkItemMessageMapper : IMessageMapper<WorkItemEventMessage>
+    public WorkItemEventMessage FromMagetMessage(MagnetMessage magnetMessage)
     {
-        public WorkItemEventMessage FromMagetMessage(MagnetMessage magnetMessage)
-        {
-            WorkItemEventMessage msg = JsonConvert.DeserializeObject<WorkItemEventMessage>(
-                magnetMessage.GetPropertyValue("Json"));
-            return msg;
-        }
+        WorkItemEventMessage msg = JsonConvert.DeserializeObject<WorkItemEventMessage>(
+            magnetMessage.GetPropertyValue("Json"));
+        return msg;
     }
 }
