+++ src/Client/Client.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AssemblyName>Magnet.Client</AssemblyName>
     <RootNamespace>Magnet.Client</RootNamespace>
     <PackageId>Magnet.Client</PackageId>
