+++ src/Client/DefaultMessageTypeRegistrations.cs
@@ -1,26 +1,22 @@
-
-
 using Magnet.Client.Mappers;
 
-namespace Magnet.Client
+namespace Magnet.Client;
+
+public static class DefaultMessageTypeRegistrations
 {
-    public static class DefaultMessageTypeRegistrations
-    {
-        public static readonly MessageTypeRegistration Email =
-             new MessageTypeRegistration
-             {
-                 Name = "Email",
-                 MessageType = typeof(EmailMessage),
-                 Mapper = new EmailMapper()
-             };
+    public static readonly MessageTypeRegistration Email =
+         new MessageTypeRegistration
+         {
+             Name = "Email",
+             MessageType = typeof(EmailMessage),
+             Mapper = new EmailMapper()
+         };
 
-        public static readonly MessageTypeRegistration Sms =
-             new MessageTypeRegistration
-             {
-                 Name = "Sms",
-                 MessageType = typeof(SmsMessage),
-                 Mapper = new SmsMapper()
-             };
-
-    }
+    public static readonly MessageTypeRegistration Sms =
+         new MessageTypeRegistration
+         {
+             Name = "Sms",
+             MessageType = typeof(SmsMessage),
+             Mapper = new SmsMapper()
+         };
 }
