+++ src/Client/IMessageStreamClient.cs
@@ -1,18 +1,17 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Magnet.Client
-{
-    public interface IMessageStreamClient
-    {
-        Task AddReceivedReceiptAsync(
-            MessageReceivedReceipt readReceipt,
-            CancellationToken cancellationToken);
+namespace Magnet.Client;
 
-        Task<MagnetMessage> GetNextAsync(
-            string clientName,
-            CancellationToken cancellationToken);
-        Task<string> Subscribe(string clientName);
-        Task UnSubscribe(string clientName);
-    }
+public interface IMessageStreamClient
+{
+    Task AddReceivedReceiptAsync(
+        MessageReceivedReceipt readReceipt,
+        CancellationToken cancellationToken);
+
+    Task<MagnetMessage> GetNextAsync(
+        string clientName,
+        CancellationToken cancellationToken);
+    Task<string> Subscribe(string clientName);
+    Task UnSubscribe(string clientName);
 }
