+++ src/Client/MagnetClient.cs
@@ -1,28 +1,27 @@
 using System.Threading.Tasks;
 using Magnet.Client.Mappers;
 
-namespace Magnet.Client
+namespace Magnet.Client;
+
+public class MagnetClient
 {
-    public class MagnetClient
+    internal MessageMapperFactory MessageMapper { get; }
+    internal IMessageStreamClient MessageStreamClient { get; }
+    internal MagnetOptions Options { get; }
+
+    public MagnetClient(
+        MessageMapperFactory messageMapper,
+        IMessageStreamClient messageStreamClient,
+        MagnetOptions options)
     {
-        internal MessageMapperFactory MessageMapper { get; }
-        internal IMessageStreamClient MessageStreamClient { get; }
-        internal MagnetOptions Options { get; }
+        MessageMapper = messageMapper;
+        MessageStreamClient = messageStreamClient;
+        Options = options;
+    }
 
-        public MagnetClient(
-            MessageMapperFactory messageMapper,
-            IMessageStreamClient messageStreamClient,
-            MagnetOptions options)
-        {
-            MessageMapper = messageMapper;
-            MessageStreamClient = messageStreamClient;
-            Options = options;
-        }
-
-        public async Task<MessageReceiver> StartAsync()
-        {
-            var queueName = await MessageStreamClient.Subscribe(Options.ClientName);
-            return new MessageReceiver(this, queueName);
-        }
+    public async Task<MessageReceiver> StartAsync()
+    {
+        var queueName = await MessageStreamClient.Subscribe(Options.ClientName);
+        return new MessageReceiver(this, queueName);
     }
 }
