+++ src/Client/MagnetClientBuilder.cs
@@ -2,47 +2,46 @@ using System;
 using Magnet.Client.Mappers;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Magnet.Client
+namespace Magnet.Client;
+
+public class MagnetClientBuilder
 {
-    public class MagnetClientBuilder
+    private readonly IServiceCollection _services;
+
+    public MagnetClientBuilder(IServiceCollection services)
     {
-        private readonly IServiceCollection _services;
+        _services = services;
+    }
 
-        public MagnetClientBuilder(IServiceCollection services)
+    public MagnetClientBuilder AddMessageType<TMessageType>(
+        string name,
+        IMessageMapper<TMessageType> mapper)
+    {
+        var typeReg = new MessageTypeRegistration
         {
-            _services = services;
-        }
+            MessageType = typeof(TMessageType),
+            Name = name,
+            Mapper = mapper
+        };
+        _services.AddSingleton(typeReg);
+        return this;
+    }
 
-        public MagnetClientBuilder AddMessageType<TMessageType>(
-            string name,
-            IMessageMapper<TMessageType> mapper)
+    public MagnetClientBuilder UseHttp(string baseUrl)
+    {
+        _services.AddSingleton<IMessageStreamClient, HttpMessageStreamClient>();
+        _services.AddHttpClient("Magnet")
+            .ConfigureHttpClient(c => c.BaseAddress = new Uri(baseUrl));
+        return this;
+    }
+
+    public MagnetClientBuilder WithClientName(string name)
+    {
+        _services.AddSingleton(new MagnetOptions
         {
-            var typeReg = new MessageTypeRegistration
-            {
-                MessageType = typeof(TMessageType),
-                Name = name,
-                Mapper = mapper
-            };
-            _services.AddSingleton(typeReg);
-            return this;
-        }
+            ClientName = name
+        });
 
-        public MagnetClientBuilder UseHttp(string baseUrl)
-        {
-            _services.AddSingleton<IMessageStreamClient, HttpMessageStreamClient>();
-            _services.AddHttpClient("Magnet")
-                .ConfigureHttpClient(c => c.BaseAddress = new Uri(baseUrl));
-            return this;
-        }
-
-        public MagnetClientBuilder WithClientName(string name)
-        {
-            _services.AddSingleton(new MagnetOptions
-            {
-                ClientName = name
-            });
-
-            return this;
-        }
+        return this;
     }
 }
