+++ src/Client/MagnetOptions.cs
@@ -1,21 +1,20 @@
-namespace Magnet.Client
+namespace Magnet.Client;
+
+public class MagnetOptions
 {
-    public class MagnetOptions
-    {
-        public string ClientName { get; set; }
+    public string ClientName { get; set; }
 
-        public GrpcOptions Grpc { get; set; }
+    public GrpcOptions Grpc { get; set; }
 
-        public SignalROptions SignalR { get; set; }
-    }
+    public SignalROptions SignalR { get; set; }
+}
 
-    public class GrpcOptions
-    {
-        public string Address { get; set; }
-    }
+public class GrpcOptions
+{
+    public string Address { get; set; }
+}
 
-    public class SignalROptions
-    {
-        public string Endpoint { get; set; }
-    }
+public class SignalROptions
+{
+    public string Endpoint { get; set; }
 }
