+++ src/Client/Mappers/IMessageMapper.cs
@@ -1,13 +1,11 @@
-namespace Magnet.Client.Mappers
+namespace Magnet.Client.Mappers;
+
+public interface IMessageMapper
 {
-    public interface IMessageMapper
-    {
 
-    }
+}
 
-    public interface IMessageMapper<TMessage> : IMessageMapper
-    {
-        //MagnetMessage ToMagnetMessage(TMessage message);
-        TMessage FromMagetMessage(MagnetMessage magnetMessage);
-    }
+public interface IMessageMapper<TMessage> : IMessageMapper
+{
+    TMessage FromMagetMessage(MagnetMessage magnetMessage);
 }
