+++ src/Client/Mappers/SmsMapper.cs
@@ -1,22 +1,21 @@
 using System.Collections.Generic;
 
-namespace Magnet.Client.Mappers
+namespace Magnet.Client.Mappers;
+
+public class SmsMapper : IMessageMapper<SmsMessage>
 {
-    public class SmsMapper : IMessageMapper<SmsMessage>
+    public SmsMessage FromMagetMessage(MagnetMessage magnetMessage)
     {
-        public SmsMessage FromMagetMessage(MagnetMessage magnetMessage)
+        var sms = new SmsMessage
         {
-            var sms = new SmsMessage
-            {
-                Id = magnetMessage.Id,
-                From = magnetMessage.From,
-                To = magnetMessage.To[0],
-                ReceivedAt = magnetMessage.ReceivedAt,
-                Body = magnetMessage.Body,
-                Properties = new Dictionary<string, string>(
-                    magnetMessage.GetPropertiesFiltered("Html", "Subject"))
-            };
-            return sms;
-        }
+            Id = magnetMessage.Id,
+            From = magnetMessage.From,
+            To = magnetMessage.To[0],
+            ReceivedAt = magnetMessage.ReceivedAt,
+            Body = magnetMessage.Body,
+            Properties = new Dictionary<string, string>(
+                magnetMessage.GetPropertiesFiltered("Html", "Subject"))
+        };
+        return sms;
     }
 }
