+++ src/Client/MessageTypeRegistration.cs
@@ -1,14 +1,13 @@
 using System;
 using Magnet.Client.Mappers;
 
-namespace Magnet.Client
+namespace Magnet.Client;
+
+public class MessageTypeRegistration
 {
-    public class MessageTypeRegistration
-    {
-        public Type MessageType { get; set; }
+    public Type MessageType { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public IMessageMapper Mapper { get; set; }
-    }
+    public IMessageMapper Mapper { get; set; }
 }
