+++ src/GraphQL/GraphQL.csproj
@@ -20,8 +20,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="HotChocolate.AspNetCore" Version="12.3.2" />
-    <PackageReference Include="HotChocolate.AspNetCore.Authorization" Version="12.3.2" />
+    <PackageReference Include="HotChocolate.AspNetCore" Version="12.4.1" />
+    <PackageReference Include="HotChocolate.AspNetCore.Authorization" Version="12.4.1" />
   </ItemGroup>
 
   <ItemGroup>
