+++ src/Host/Startup.cs
@@ -1,5 +1,9 @@
+using System.Collections.Generic;
+using System.Security.Claims;
+using System.Threading.Tasks;
 using Magnet.GraphQL;
 using Magnet.Store.Mongo;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
@@ -26,6 +30,7 @@ public class Startup
         {
             options.AllowSynchronousIO = true;
         });
+
         services.AddControllers();
         services.AddMagnet()
                     .AddSendGridEmail()
@@ -33,6 +38,9 @@ public class Startup
                     .AddAzureDevOps()
                     .AddRabbitMQ(Configuration)
                     .AddMongoStore(Configuration);
+
+        services.AddAuthentication("fake")
+            .AddCookie("fake");
 
         services.AddAuthorization(o => o.AddPolicy(
             "Magnet.Read",
@@ -64,12 +72,19 @@ public class Startup
         app.UseRouting();
 
         app.UseAuthorization();
+        app.UseAuthentication();
+
+        app.Use(async (context, next) =>
+        {
+            var identity = new ClaimsIdentity( new List<Claim>() { new Claim("sub", "admin") }, "fake");
+            await context.SignInAsync("fake", new ClaimsPrincipal(identity));
+            await next.Invoke();
+        });
 
         app.UseEndpoints(endpoints =>
         {
             endpoints.MapControllers();
             endpoints.MapGraphQL();
         });
-
     }
 }
