+++ src/Messaging.AzureServiceBus/Messaging.AzureServiceBus.csproj
@@ -18,7 +18,6 @@
     <DebugSymbols>true</DebugSymbols>
   </PropertyGroup>
 
-
   <ItemGroup>
     <PackageReference Include="Microsoft.Azure.ServiceBus" Version="4.1.1" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
