+++ src/Models/MagnetMessage.cs
@@ -1,23 +1,22 @@
 using System;
 using System.Collections.Generic;
 
-namespace Magnet
+namespace Magnet;
+
+public class MagnetMessage
 {
-    public class MagnetMessage
-    {
-        public Guid Id { get; set; }
+    public Guid Id { get; set; }
 
-        public string Type { get; set; }
+    public string Type { get; set; }
 
-        public DateTime ReceivedAt { get; set; }
+    public DateTime ReceivedAt { get; set; }
 
-        public string From { get; set; }
+    public string From { get; set; }
 
-        public IList<string> To { get; set; }
+    public IList<string> To { get; set; }
 
-        public string Body { get; set; }
+    public string Body { get; set; }
 
-        public IReadOnlyDictionary<string, string> Properties { get; set; }
-        public string Provider { get; set; }
-    }
+    public IReadOnlyDictionary<string, string> Properties { get; set; }
+    public string Provider { get; set; }
 }
