+++ src/Models/MessageReceivedReceipt.cs
@@ -1,15 +1,14 @@
 using System;
 
-namespace Magnet
+namespace Magnet;
+
+public class MessageReceivedReceipt
 {
-    public class MessageReceivedReceipt
-    {
-        public Guid MessageId { get; set; }
+    public Guid MessageId { get; set; }
 
-        public bool IsMatch { get; set; }
+    public bool IsMatch { get; set; }
 
-        public string ClientName { get; set; }
+    public string ClientName { get; set; }
 
-        public DateTime ReceivedAt { get; set; }
-    }
+    public DateTime ReceivedAt { get; set; }
 }
