+++ src/Models/Models.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AssemblyName>Magnet.Models</AssemblyName>
     <RootNamespace>Magnet</RootNamespace>
     <PackageId>Magnet.Models</PackageId>
