+++ src/SendGridEmail/SendGrid.Email.csproj
@@ -21,7 +21,7 @@
 
   <ItemGroup>
       <FrameworkReference Include="Microsoft.AspNetCore.App" />
-      <PackageReference Include="StrongGrid" Version="0.61.0">
+      <PackageReference Include="StrongGrid" Version="0.84.0">
         <PrivateAssets>all</PrivateAssets>
       </PackageReference>
   </ItemGroup>
